+++ TwitterJSON/TJUsers.swift
@@ -9,23 +9,12 @@
 import Foundation
 import SwiftyJSON
 
-public protocol TJUsersDelegate {
-    /**
-    Delegate method which contains an array of TJUser objets
-    
-    :param: Array Collection of TJUser objets.
-    */
-    func gotUsers(users:[TJUser])
-}
-
 public class TJUsers {
     
     private let twitterJSON: TwitterJSON!
-    public var delegate: TJUsersDelegate?
     
     public init(apiKey: String, apiSecret: String) {
         self.twitterJSON = TwitterJSON(apiKey: apiKey, apiSecret: apiSecret)
-        self.twitterJSON.delegate = self
     }
     
     /**
@@ -34,26 +23,18 @@ public class TJUsers {
     
     :param: String Screen name of the users whos followers to retrieve.
     */
-    public func getFollowersForUser(screenName: String) {
+    public func getFollowersForUser(screenName: String, completion: (users: [TJUser]) -> Void) {
         twitterJSON.getBearerToken { (bearerToken) -> Void in
             let apiURL = "https://api.twitter.com/1.1/followers/list.json?screen_name=" + screenName
             self.twitterJSON.performDataRequestForURL(apiURL, bearerToken: bearerToken, completion: { data in
-            
+                var users = [TJUser]()
+                for item in data["users"] {
+                    let user = TJUser(userInfo: item.1)
+                    users.append(user)
+                }
+                completion(users: users)
             })
         }
     }
 
-}
-
-extension TJUsers: TwitterJSONDelegate {
-    
-    public func gotdata(data: JSON) {
-        var users = [TJUser]()
-        for item in data["users"] {
-            let user = TJUser(userInfo: item.1)
-            users.append(user)
-        }
-        self.delegate?.gotUsers(users)
-    }
-    
 }